Fixed Box2D borders, sprites sizes

+++ IlluminatiConfirmed/Game/constants.h
@@ -6,7 +6,7 @@
 const int WINDOW_HEIGHT = 800;
 const int WINDOW_WIDTH = 800;
 
-const int SCALE = 32;
+const float SCALE = 32.0;
 
 enum class Direction { Up, Down, Left, Right };
 
@@ -22,5 +22,5 @@ sf::Vector2<T> B2Vec2toSfVector2(const b2Vec2 &vector) {
 
 template <typename V1, typename V2>
 float RadBetweenVectors(const V1 &v1, const V2 &v2) {
-  return static_cast<float>(atan2(v2.y - v1.y), v2.x - v1.x);
+  return static_cast<float>(atan2(v2.y - v1.y, v2.x - v1.x));
 }
